requirements.txt
asgiref==3.9.2
Django==5.2.6
django-crispy-forms==2.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
